apply refactoring PR comments
